rename iac sidebar file
